[DFB]: Move host-side DFB files into the correct experimental dir
